Checkpoint 23 Completed
Added a UIAlertAction that happens every time the app is loaded or
refreshes. Also added my own extra: if there is no “.” in the
textField, the browser will automatically do a google search. So if
someone types in no spaces and a “.” the browser will look for the
website. But any space (how to do X) or lack of a “.” (apple) will go
straight to google.