Checkpoint 24 Complete
Added all of the code necessary to make a custom
JLAwesomeFloatingToolbar UIView. Additionally, I used #define to
refactor some NSLocalizedStrings, allowing it to more easily be
converted to other languages. Finally, I figured out how to make the
custom toolbar go to the center/bottom of the screen, no matter the
orientation.