Checkpoint 25 Still Not Complete
Hi Mark- could you take a look at the didtrytopinch method. I’ve gotten
it to the point where my app knows I’m pinching on the screen, but the
toolbar isn’t doing what I want it to. I’ve been working on this for a
few hours and think I need a second pair of eyes. Thanks.